Add Day 2, p2
Had to pivot from my bfs idea to an other method. Problem with the bfs
with bitmask idea was that there are too many pcs to have them all be
represented as a unique bitmask. Due to the number of pcs, the bitmask
of the last pc converted to an integer was higher than PHP_INT_MAX,
which caused issues.

Now I use bitmask to generate the possible LAN configuration for each
network, and then check if it's a valid LAN. Starting from the biggest
possible LAN (all 1's) and then checking all variants where one pc is
removed, then 2, etc. We stop checking if the number of pcs gets lower
than the current biggest LAN network, because we obviously won't find a
bigger LAN in this network then.